requirements.txt
matplotlib 
pandas 
pillow 
jupyter
